+++ src/handlers/license_manager_setting_handler.py
@@ -29,6 +29,7 @@ from validators.swagger_request_models import (
     LicenseManagerClientSettingPostModel,
     LicenseManagerConfigSettingPostModel,
     LicenseManagerConfigSettingPatchModel,
+    LicenseManagerClientSettingPatchModel,
 )
 from validators.utils import validate_kwargs 
 
@@ -55,6 +56,7 @@ class LicenseManagerClientHandler(AbstractHandler):
             Endpoint.SETTINGS_LICENSE_MANAGER_CLIENT: {
                 HTTPMethod.GET: self.get,
                 HTTPMethod.POST: self.post,
+                HTTPMethod.PATCH: self.patch,
                 HTTPMethod.DELETE: self.delete,
             }
         }
@@ -142,6 +144,85 @@ class LicenseManagerClientHandler(AbstractHandler):
                 alg=alg,
                 kid=kid,
                 public_key=self.get_public(prk)
+            )
+        )
+
+    @validate_kwargs
+    def patch(
+        self,
+        event: LicenseManagerClientSettingPatchModel,
+    ):
+        new_kid = event.key_id
+        new_alg = event.algorithm
+        new_raw_prk: str = event.private_key
+
+        setting = self.settings_service. \
+            get_license_manager_client_key_data(value=False)
+
+        if not setting:
+            return build_response(
+                code=HTTPStatus.NOT_FOUND,
+                content='License Manager Client-Key does not exist.'
+            )
+
+        configuration = setting.value
+        kid = configuration.get(KID_ATTR)
+        alg = configuration.get(ALG_ATTR)
+
+        if not (kid and alg):
+            msg = (
+                "License Manager Client-Key does not contain 'kid' or 'alg' data."
+            )
+            _LOG.warning(msg)
+            return build_response(
+                code=HTTPStatus.NOT_FOUND,
+                content=msg,
+            )
+
+        try:
+            new_prk = load_pem_private_key(new_raw_prk.encode(), None)
+        except (ValueError, Exception):
+            raise ResponseFactory(HTTPStatus.BAD_REQUEST).message(
+                'Invalid private key'
+            ).exc()
+
+        if new_kid is not None and new_kid != kid:
+            name_to_del = LmTokenProducer.derive_client_private_key_id(kid=kid)
+            name = LmTokenProducer.derive_client_private_key_id(kid=new_kid)
+            kid = new_kid
+        else:
+            name = name_to_del = LmTokenProducer. \
+                derive_client_private_key_id(kid=kid)
+
+        _LOG.debug(f"Deleting old private key with name '{name_to_del}'")
+        self._ssm_client.delete_parameter(name_to_del)
+
+        self._ssm_client.create_secret(
+            secret_name=name,
+            secret_value={
+                VALUE_ATTR: new_prk.private_bytes(
+                    Encoding.PEM,
+                    PrivateFormat.PKCS8,
+                    NoEncryption()
+                ).decode('utf-8')
+            }
+        )
+        _LOG.info('Private key was saved to SSM')
+
+        # Update existing configuration
+        configuration[KID_ATTR] = kid
+        configuration[ALG_ATTR] = new_alg
+        setting.value = configuration
+
+        _LOG.info(f'Updating License Manager Client-Key data: {setting.value}.')
+        self.settings_service.save(setting=setting)
+
+        return build_response(
+            code=HTTPStatus.OK,
+            content=self.get_dto(
+                alg=new_alg,
+                kid=kid,
+                public_key=self.get_public(new_prk)
             )
         )