+++ src/helpers/constants.py
@@ -887,6 +887,7 @@ class Permission(str, Enum):
     SETTINGS_DELETE_LM_CONFIG = 'settings:delete_lm_config', True
     SETTINGS_DESCRIBE_LM_CLIENT = 'settings:describe_lm_client'
     SETTINGS_CREATE_LM_CLIENT = 'settings:create_lm_client', True
+    SETTINGS_UPDATE_LM_CLIENT = 'settings:update_lm_client', True
     SETTINGS_DELETE_LM_CLIENT = 'settings:delete_lm_client', True
 
     RABBITMQ_DESCRIBE = 'rabbitmq:describe'