+++ src/validators/registry.py
@@ -94,6 +94,7 @@ from validators.swagger_request_models import (
     ResourcesExceptionsGetModel,
     ResourcesExceptionsPostModel,
     LicenseManagerConfigSettingPatchModel,
+    LicenseManagerClientSettingPatchModel,
 )
 from validators.swagger_response_models import (
     CredentialsActivationModel,
@@ -969,6 +970,15 @@ data: tuple[EndpointInfo, ...] = (
         responses=[(HTTPStatus.CREATED, SingleLMClientModel, None)],
         permission=Permission.SETTINGS_CREATE_LM_CLIENT,
         description='Allows to add license manager client'
+    ),
+    EndpointInfo(
+        path=Endpoint.SETTINGS_LICENSE_MANAGER_CLIENT,
+        method=HTTPMethod.PATCH,
+        lambda_name=LambdaName.CONFIGURATION_API_HANDLER,
+        request_model=LicenseManagerClientSettingPatchModel,
+        responses=[(HTTPStatus.OK, SingleLMClientModel, None)],
+        permission=Permission.SETTINGS_UPDATE_LM_CLIENT,
+        description='Allows to update license manager client'
     ),
     EndpointInfo(
         path=Endpoint.SETTINGS_LICENSE_MANAGER_CLIENT,